feat(neo4j): because the db queries are longer than other adapters. queries are grouped by type
