+++ packages/neo4j/src/utils.ts
@@ -18,41 +18,3 @@ export const neo4jEpochToDate = (epoch: typeof neo4j.Integer) => {
 
   return new Date(epochParsed)
 }
-
-export const userReturn = `
-  { 
-    id: u.id,
-    name: u.name,
-    email: u.email,
-    image: u.image,
-    emailVerified: u.emailVerified.epochMillis
-  } AS user 
-`
-export const sessionReturn = `
-  {
-    userId: u.id,
-    id: s.id,
-    expires: s.expires.epochMillis, 
-    accessToken: s.accessToken,
-    sessionToken: s.sessionToken
-  } AS session
-`
-export const accountReturn = `
-  {
-    userId: u.id,
-    providerId: a.providerId,
-    providerAccountId: a.providerAccountId,
-    providerType: a.providerType,
-    refreshToken: a.refreshToken,
-    accessToken: a.accessToken,
-    accessTokenExpires: a.accessTokenExpires
-  } AS account
-`
-
-export const verificationRequestReturn = `
-  {
-    identifier: v.identifier,
-    token: v.token,
-    expires: v.expires.epochMillis
-  } AS verificationRequest 
-`
